+++ tests/common.py
@@ -12,7 +12,8 @@ import pathlib
 from mongoengine import connect, disconnect, connection
 
 from src.features.smarterdb import (
-    DB_ALIAS, Breed, BreedAlias, Counter, Dataset, SampleSheep, VariantSheep)
+    DB_ALIAS, Breed, BreedAlias, Counter, Dataset, SampleSheep, VariantSheep,
+    IlluminaChip)
 
 FIXTURES_DIR = pathlib.Path(__file__).parent / "fixtures"
 
@@ -128,5 +129,21 @@ class VariantsMixin():
     @classmethod
     def tearDownClass(cls):
         VariantSheep.objects.delete()
+
+        super().tearDownClass()
+
+
+class IlluminaChipMixin():
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+
+        cls.chip_name = "IlluminaOvineSNP50"
+        cls.chip = IlluminaChip(name=cls.chip_name, species="Sheep")
+        cls.chip.save()
+
+    @classmethod
+    def tearDownClass(cls):
+        IlluminaChip.objects.delete()
 
         super().tearDownClass()